Restructured the code and altered the test phase
