+++ .circleci/config.yml
@@ -12,7 +12,7 @@ jobs:
       - python/install-deps
       - python/save-cache
       - run:
-          command: ./manage.py test
+          command: python -m unittest discover
           name: Test
 
 workflows:
